Split out evaluator helpers from translate.py
